Move OSInfo to top level package an make it package private
